% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_utils.R
\name{.convert_inchikey_to_inchi_pubchem}
\alias{.convert_inchikey_to_inchi_pubchem}
\title{Get an InChI value from an InChIKey via Pubchem}
\usage{
.convert_inchikey_to_inchi_pubchem(inchikey)
}
\arguments{
\item{inchikey}{An InChIKey.}
}
\value{
character, InChI.
}
\description{
Uses the PubChem PUG REST API, this service does not require an API key. I think it will always provide a standard InChI?
}
