% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_chembl.R
\name{.format_chembl_names}
\alias{.format_chembl_names}
\title{Organize chembl dataframes.}
\usage{
.format_chembl_names(names_df)
}
\value{
TBD
}
\description{
Organize chembl data and export in DTEX-mergeable format
}
