% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_chembl.R
\name{.format_chembl_structures}
\alias{.format_chembl_structures}
\title{Organize chembl structures}
\usage{
.format_chembl_structures(activity_df, structures_df, names_df)
}
\value{
TBD
}
\description{
Organize chembl data and export in DTEX-mergeable format
}
