% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dgidb.R
\name{.format_dgidb_structures}
\alias{.format_dgidb_structures}
\title{Placeholder}
\usage{
.format_dgidb_structures(
  dgidb_df,
  path_to_inchi_tsv,
  path_to_inchikey_tsv,
  path_to_smiles_tsv
)
}
\value{
TBD
}
\description{
Placeholder
}
