% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_drugbank.R
\name{.format_drugbank_activities}
\alias{.format_drugbank_activities}
\title{Placeholder}
\usage{
.format_drugbank_activities(drugbank_db, drugbank_structures)
}
\value{
TBD
}
\description{
Placeholder
}
