% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_drugbank.R
\name{.format_drugbank_names}
\alias{.format_drugbank_names}
\title{Export the DrugBank names.}
\usage{
.format_drugbank_names(drugbank_structures)
}
\value{
TBD
}
\description{
Export the DrugBank names.
}
