% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_chembl.R
\name{.get_chembl_files}
\alias{.get_chembl_files}
\title{Retrieve chembl data from TSVs.}
\usage{
.get_chembl_files(activity_file_id, names_file_id, structures_file_id)
}
\value{
TBD
}
\description{
Retrieve chembl data from TSVs extracted from the mysql database. Files should be hosted on synapse.
}
