% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_chemicalprobes.R
\name{.get_chemicalprobes_data}
\alias{.get_chemicalprobes_data}
\title{Retrieve chemicalprobes data from CSVs.}
\usage{
.get_chemicalprobes_data(synapse_id)
}
\arguments{
\item{synapse_id}{The Synapse ID of the ChemicalProbes.org exported csv.}
}
\value{
TBD
}
\description{
Retrieve chemicalprobes data from CSV (db exports must be requested by email as of early 2021). Files should be hosted on synapse.
}
