% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_drugbank.R
\name{.get_drugbank_data}
\alias{.get_drugbank_data}
\title{Retrieve drugbank data from database.}
\usage{
.get_drugbank_data(db_synapse_id, structures_synapse_id)
}
\arguments{
\item{db_synapse_id}{The Synapse ID of the drugbank XML.}

\item{structures_synapse_id}{The Synapse ID of the structures CSV.}
}
\value{
TBD
}
\description{
Retrieve drugbank data from exported XML and csvs. Files should be hosted on synapse.
}
