% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dgidb.R
\name{prepare_dgidb_data}
\alias{prepare_dgidb_data}
\title{Prepare dgidb drugs for structure mapping}
\usage{
prepare_dgidb_data(
  dgidb_interactions_url,
  omit_sources = c("DrugBank", "ChemblInteractions")
)
}
\arguments{
\item{dgidb_interactions_url}{A url of the drug gene interaction database, eg "https://www.dgidb.org/data/monthly_tsvs/2021-Jan/interactions.tsv"}
}
\value{
An input tsv for the PubChem Identifier Exchange Service
}
\description{
Generates a formatted CSV that can be used as input into the PubChem Identifier Exchange Service (Input ID list = synonyms, operator type = same CID, Output IDs = InChIs,InChiKey,smiles)
}
