% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_chembl.R
\name{process_chembl}
\alias{process_chembl}
\title{Organize chembl dataframes.}
\usage{
process_chembl(activity_file_id, names_file_id, structures_file_id)
}
\value{
TBD
}
\description{
Organize chembl data and export in DTEX-mergeable format
}
