% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dgidb.R
\name{.format_dgidb_structures}
\alias{.format_dgidb_structures}
\title{This is a a draft function.}
\usage{
.format_dgidb_structures(
  dgidb_df,
  path_to_inchi_tsv,
  path_to_inchikey_tsv,
  path_to_smiles_tsv
)
}
\value{
TBD
Placeholder

TBD
}
\description{
Placeholder

Placeholder
}
